ui fix

+++ Foodorderingsystem/service/impl/FoodItemServiceImpl.java
@@ -1,6 +1,7 @@
 package Foodorderingsystem.service.impl;
 
 import Foodorderingsystem.model.FoodItem;
+import Foodorderingsystem.model.User;
 import Foodorderingsystem.repository.impl.FoodItemRepositoryImpl;
 import Foodorderingsystem.service.FoodItemService;
 
@@ -17,15 +18,26 @@ public class FoodItemServiceImpl implements FoodItemService {
         return instance;
     }
     static int id=0;
+    UserServiceImpl userService=UserServiceImpl.getInstance();
     @Override
-    public FoodItem addFoodItem( String restaurantId, String name, String description, String price, boolean availabilty) {
+    public FoodItem addFoodItem( String ownerId,String restaurantId, String name, String description, String price, boolean availabilty) {
+        User user=userService.getUserByUserId(ownerId);
+        if(user== null || user.getRole()!="OWNER")
+        {
+            return null;
+        }
         FoodItem foodItem=new FoodItem("food"+(++id),restaurantId,name,description,price,availabilty);
         foodItemRepository.saveFoodItem(foodItem);
         return foodItem;
     }
 
     @Override
-    public FoodItem updateFoodItem(String foodItemId, String price,boolean availabilty) {
+    public FoodItem updateFoodItem(String ownerId,String foodItemId, String price,boolean availabilty) {
+        User user=userService.getUserByUserId(ownerId);
+        if(user== null || user.getRole()!="OWNER")
+        {
+            return null;
+        }
         FoodItem foodItem= foodItemRepository.findByFoodItemId(foodItemId);
         if(foodItem!=null)
         {
@@ -38,7 +50,12 @@ public class FoodItemServiceImpl implements FoodItemService {
     }
 
     @Override
-    public FoodItem deleteFoodItem(String foodItemId) {
+    public FoodItem deleteFoodItem(String ownerId,String foodItemId) {
+        User user=userService.getUserByUserId(ownerId);
+        if(user== null || user.getRole()!="OWNER")
+        {
+            return null;
+        }
         FoodItem foodItem=foodItemRepository.removeFoodItemId(foodItemId);
         if(foodItem!=null)
         {
